modified:   src/MechanicalSketch.jl empy_figure include height width modified:   src/colorlegends.jl modified:   src/colorlegends_vector.jl Separate out draw_background modified:   src/colors.jl      Add chroma, saturation, color_with_saturation modified:   src/curves.jl      Accept quantities that scales to length. +rect. modified:   src/place_image.jl Scale pngs, improve modify_latex modified:   src/scale.jl       get_scale_sketch::Float64 modified:   src/table.jl       Document pixelwidth (WIP) modified:   test/test_1.jl     Comment modified:   test/test_10.png   Installed font with lightning symbol modified:   test/test_11.jl    Add missing import modified:   test/test_11.png   Update with coloured latex modified:   test/test_13.jl    Color change modified:   test/test_13.png   Color change modified:   test/test_19.jl    Test png scale modified:   test/test_19.png   Test png scale modified:   test/test_20.jl    Add missing import modified:   test/test_36.jl    WIP modified:   test/test_36.png   WIP new file:   test/test_functions_36.jl WIP

+++ src/scale.jl
@@ -58,11 +58,11 @@ function set_scale_sketch()
 end
 
 """
-    get_scale_sketch(q::Quantity)
+    get_scale_sketch(q::Quantity)::Float64
 
 Return the number of pixels (or points, this may be ambiguous) corresponding to a given quantity.
 
-    get_scale_sketch(u::FreeUnits)
+    get_scale_sketch(u::FreeUnits)::Float64
 
 Return the number of pixels (or points, this may be ambiguous) corresponding to one unit of u.
 
@@ -74,11 +74,11 @@ Return the number of pixels (or points, this may be ambiguous) corresponding to 
     set_scale_sketch()
     70 m/s
 """
-get_scale_sketch(q::Length) = upreferred(q / SCALEDIST)::Float64
-get_scale_sketch(q::Velocity) = upreferred(q / SCALEVELOCITY)::Float64
-get_scale_sketch(q::Force) = upreferred(q / SCALEFORCE)::Float64
-get_scale_sketch(u::FreeUnits) = get_scale_sketch(1∙u)::Float64
-get_scale_sketch(x) = x
+get_scale_sketch(q::Length)::Float64 = upreferred(q / SCALEDIST)
+get_scale_sketch(q::Velocity)::Float64 = upreferred(q / SCALEVELOCITY)
+get_scale_sketch(q::Force)::Float64 = upreferred(q / SCALEFORCE)
+get_scale_sketch(u::FreeUnits)::Float64 = get_scale_sketch(1∙u)
+get_scale_sketch(x)::Float64 = x
 """
     set_figure_height(h::Int)
